Skip d-m-p if called in top-level pom.xml
In order to allow a `docker:start` from the top-level so that it descents into the submodules, the execution of d-m-p in the toplevel pom.xml must be switched off with `<skip>true</skip>`. Unfortunately since the submodules inherit from the parent pom.xml, the plugin needs to be reenabled in the child poms again, leading to some duplications (unfortunately I couldn't realize a simple solution involving submodules).

An alternative would be to use a single Maven project for the Docker builds, where all images are configured within a single d-m-p config (within the <images> list).

 BTW, for the setup described above (single project) there will be direct docker compose support soon (i.e. d-m-p picking up a docker compose configuration as well as creating docker compose files from the d-m-p configuration).